requirements.txt
streamlit
sentence-transformers
faiss-cpu
PyPDF2
groq
python-dotenv
uuid
